+++ libraries/transformer/src/main/java/androidx/media3/transformer/ExoAssetLoaderVideoRenderer.java
@@ -25,6 +25,7 @@ import androidx.media3.common.C;
 import androidx.media3.common.ColorInfo;
 import androidx.media3.common.Format;
 import androidx.media3.decoder.DecoderInputBuffer;
+import androidx.media3.exoplayer.source.SampleStream;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
@@ -165,7 +166,7 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
   protected void onDecoderInputReady(DecoderInputBuffer inputBuffer) {
     long durationUs = getPeriodDurationUs();
     boolean exceedsStrictDuration =
-        isPeriodDurationStrict()
+        ((checkNotNull(getStream()).getFlags() & SampleStream.FLAG_STRICT_DURATION) != 0)
             && durationUs != C.TIME_UNSET
             && (inputBuffer.timeUs - getStreamOffsetUs()) >= durationUs;
     if (inputBuffer.timeUs < getLastResetPositionUs() || exceedsStrictDuration) {